+++ tests/Unit/Model/BlockTest.php
@@ -38,12 +38,12 @@ class BlockTest extends TestCase
                     "time_lock"        => 608971,
                     "from_length"      => 1,
                     "to_length"        => 116,
-                    "fee"              => 0,
-                    "fee_raw"          => 0,
+                    "fee"              => 8.178,
+                    "fee_raw"          => 81780,
                     "amount"           => 1000,
                     "amount_raw"       => 10000000,
-                    "total_amount"     => 1000,
-                    "total_amount_raw" => 10000000,
+                    "total_amount"     => 1008.178,
+                    "total_amount_raw" => 10081780,
                     "timestamp"        => "Wed, 20 Feb 2019 19:01:25 GMT",
                     "timestamp_UTC"    => 1550689285,
                     "timestamp_block"  => 25946973,
@@ -56,15 +56,15 @@ class BlockTest extends TestCase
                                 "trx_from_address"    => 'WEBD$gCsh0nNrsZv9VYQfe5Jn$9YMnD4hdyx62n$',
                                 "trx_from_pub_key"    => "c20d135a40514b06321e55eca97549bf5f8780cf64c225a97bb339c14aa47df6",
                                 "trx_from_signature"  => "e9ff2ac208c28845d40340653ac1b9ab66fe84817849692e2dae4057aeb25198bab054fa38659596732358ab22bad3aafc4f7794b22cf7adc60f75813bc33a01",
-                                "trx_from_amount"     => 1000,
-                                "trx_from_amount_raw" => 10000000
+                                "trx_from_amount"     => 1008.178,
+                                "trx_from_amount_raw" => 10081780
                             ]
                         ],
                         "addresses" => [
                             'WEBD$gCsh0nNrsZv9VYQfe5Jn$9YMnD4hdyx62n$'
                         ],
-                        "amount"     => 1000,
-                        "amount_raw" => 10000000
+                        "amount"     => 1008.178,
+                        "amount_raw" => 10081780
                     ],
                     "to" => [
                         "trxs" => [
@@ -117,6 +117,11 @@ class BlockTest extends TestCase
         static::assertEquals('2258804812b60576d8aa5ef8d99e1db50d609b198e2fff7adaab85cb9513095afd8ef977eed98007feee72fe7783e56656ca8c859542b40790baa6a2d1b9800d', $oBlock->getPOSSignature());
         static::assertTrue($oBlock->isPOS());
         static::assertFalse($oBlock->isPOW());
+
+        static::assertEquals(8.178, $oBlock->getFeeReward());
+        static::assertEquals(81780, $oBlock->getFeeRewardRaw());
+        static::assertEquals(6008.178, $oBlock->getTotalReward());
+        static::assertEquals(60081780, $oBlock->getTotalRewardRaw());
 
         static::assertInstanceOf(Account::class, $oBlock->getMinerAccount());
         static::assertInstanceOf(Account::class, $oBlock->getResolverAccount());