+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [v1.3.0] - 2017-02-26
+### Added
+- Block::getFeeReward
+- Block::getFeeRewardRaw
+- Block::getTotalReward
+- Block::getTotalRewardRaw
+- Transaction::isVirtual
+
 ## [v1.2.0] - 2017-02-25
 ### Added
 - getRawData for single models